+++ packages/terra-date-picker/tests/jest/DatePicker.test.jsx
@@ -66,16 +66,6 @@ it('should render a date picker with onSelect', () => {
   expect(datePicker).toMatchSnapshot();
 });
 
-it('throws error on missing locale prop in Base', () => {
-  global.console = { error: jest.fn() };
-
-  expect(() => {
-    render(<DatePicker name="date-input" />);
-  }).toThrowError();
-  // eslint-disable-next-line no-console
-  expect(console.error).toBeCalledWith(expect.stringContaining('Component is internationalized, and must be wrapped in terra-base'));
-});
-
 it('should render a date picker with onCalendarButtonClick', () => {
   const handleOnCalendarButtonClick = () => {};
   const datePicker = shallowWithIntl(<DatePicker name="date-input" onCalendarButtonClick={handleOnCalendarButtonClick} utcOffset={0} />);