+++ packages/terra-time-input/src/TimeInput.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import Input from 'terra-form-input';
 import ButtonGroup from 'terra-button-group';
+import { injectIntl, intlShape } from 'react-intl';
 
 import KeyCode from 'keycode-js';
 import TimeUtil from './TimeUtil';
@@ -25,6 +26,10 @@ const propTypes = {
    */
   // eslint-disable-next-line react/forbid-prop-types
   hourAttributes: PropTypes.object,
+  /**
+   * intl object programmatically imported through injectIntl from react-intl.
+   * */
+  intl: intlShape.isRequired,
   /**
    * Custom input attributes to apply to the minutes input
    */
@@ -75,17 +80,8 @@ const defaultProps = {
   variant: TimeUtil.FORMAT_24_HOUR,
 };
 
-const contextTypes = {
-  /* eslint-disable consistent-return */
-  intl: (context) => {
-    if (context.intl === undefined) {
-      return new Error('Component is internationalized, and must be wrapped in terra-base');
-    }
-  },
-};
-
 class TimeInput extends React.Component {
-  constructor(props, context) {
+  constructor(props) {
     super(props);
 
     let { value } = this.props;
@@ -123,7 +119,7 @@ class TimeInput extends React.Component {
     let meridiem;
 
     if (props.variant === TimeUtil.FORMAT_12_HOUR) {
-      if (!context.intl.messages['Terra.timeInput.am'] || !context.intl.messages['Terra.timeInput.pm']) {
+      if (!this.props.intl.messages['Terra.timeInput.am'] || !this.props.intl.messages['Terra.timeInput.pm']) {
         if (process.env !== 'production') {
           // eslint-disable-next-line no-console
           console.warn('This locale only uses 24 hour clock. The ante meridiem and post meridiem will not be displayed');
@@ -132,8 +128,8 @@ class TimeInput extends React.Component {
         this.anteMeridiem = '';
         this.postMeridiem = '';
       } else {
-        this.anteMeridiem = context.intl.formatMessage({ id: 'Terra.timeInput.am' });
-        this.postMeridiem = context.intl.formatMessage({ id: 'Terra.timeInput.pm' });
+        this.anteMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.am' });
+        this.postMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.pm' });
       }
 
       if (hour) {
@@ -169,7 +165,7 @@ class TimeInput extends React.Component {
     let { meridiem } = this.state;
 
     if (this.props.variant === TimeUtil.FORMAT_12_HOUR) {
-      if (!this.context.intl.messages['Terra.timeInput.am'] || !this.context.intl.messages['Terra.timeInput.pm']) {
+      if (!this.props.intl.messages['Terra.timeInput.am'] || !this.props.intl.messages['Terra.timeInput.pm']) {
         if (process.env !== 'production') {
           // eslint-disable-next-line no-console
           console.warn('This locale only uses 24 hour clock. The ante meridiem and post meridiem will not be displayed');
@@ -178,8 +174,8 @@ class TimeInput extends React.Component {
         this.anteMeridiem = '';
         this.postMeridiem = '';
       } else {
-        this.anteMeridiem = this.context.intl.formatMessage({ id: 'Terra.timeInput.am' });
-        this.postMeridiem = this.context.intl.formatMessage({ id: 'Terra.timeInput.pm' });
+        this.anteMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.am' });
+        this.postMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.pm' });
       }
 
       if (hour) {
@@ -541,6 +537,7 @@ class TimeInput extends React.Component {
       inputAttributes,
       minuteAttributes,
       hourAttributes,
+      intl,
       onBlur,
       onChange,
       onFocus,
@@ -599,7 +596,7 @@ class TimeInput extends React.Component {
             className={cx('time-input-hour')}
             value={this.state.hour}
             name={'terra-time-hour-'.concat(name)}
-            placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.hh' })}
+            placeholder={intl.formatMessage({ id: 'Terra.timeInput.hh' })}
             maxLength="2"
             onChange={this.handleHourChange}
             onKeyDown={this.handleHourInputKeyDown}
@@ -610,7 +607,7 @@ class TimeInput extends React.Component {
             disabled={disabled}
           />
           <label htmlFor={instanceHoursAttrs.id} className={cx('mobile-input-label')}>
-            {this.context.intl.formatMessage({ id: 'Terra.timeInput.hours' })}
+            {intl.formatMessage({ id: 'Terra.timeInput.hours' })}
           </label>
         </div>
         <span className={cx('time-spacer')}>:</span>
@@ -622,7 +619,7 @@ class TimeInput extends React.Component {
             className={cx('time-input-minute')}
             value={this.state.minute}
             name={'terra-time-minute-'.concat(name)}
-            placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.mm' })}
+            placeholder={intl.formatMessage({ id: 'Terra.timeInput.mm' })}
             maxLength="2"
             onChange={this.handleMinuteChange}
             onKeyDown={this.handleMinuteInputKeyDown}
@@ -633,7 +630,7 @@ class TimeInput extends React.Component {
             disabled={disabled}
           />
           <label htmlFor={instanceMinuteAttrs.id} className={cx('mobile-input-label')}>
-            {this.context.intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
+            {intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
           </label>
         </div>
         {this.props.variant === TimeUtil.FORMAT_12_HOUR && (
@@ -670,6 +667,7 @@ class TimeInput extends React.Component {
       inputAttributes,
       minuteAttributes,
       hourAttributes,
+      intl,
       onBlur,
       onChange,
       onFocus,
@@ -717,7 +715,7 @@ class TimeInput extends React.Component {
         <Input
           {...inputAttributes}
           {...minuteAttributes}
-          aria-label={this.context.intl.formatMessage({ id: 'Terra.timeInput.hours' })}
+          aria-label={intl.formatMessage({ id: 'Terra.timeInput.hours' })}
           refCallback={(inputRef) => {
             this.hourInput = inputRef;
             if (refCallback) refCallback(inputRef);
@@ -726,7 +724,7 @@ class TimeInput extends React.Component {
           type="text"
           value={this.state.hour}
           name={'terra-time-hour-'.concat(name)}
-          placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.hh' })}
+          placeholder={intl.formatMessage({ id: 'Terra.timeInput.hh' })}
           maxLength="2"
           onChange={this.handleHourChange}
           onKeyDown={this.handleHourInputKeyDown}
@@ -741,12 +739,12 @@ class TimeInput extends React.Component {
           {...inputAttributes}
           {...minuteAttributes}
           refCallback={(inputRef) => { this.minuteInput = inputRef; }}
-          aria-label={this.context.intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
+          aria-label={intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
           className={cx('time-input-minute', 'desktop', { 'initial-focus': this.state.minuteInitialFocused })}
           type="text"
           value={this.state.minute}
           name={'terra-time-minute-'.concat(name)}
-          placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.mm' })}
+          placeholder={intl.formatMessage({ id: 'Terra.timeInput.mm' })}
           maxLength="2"
           onChange={this.handleMinuteChange}
           onKeyDown={this.handleMinuteInputKeyDown}
@@ -760,7 +758,7 @@ class TimeInput extends React.Component {
           [
             <Input
               {...inputAttributes}
-              aria-label={this.context.intl.formatMessage({ id: 'Terra.timeInput.display.meridiem' })} // value in translations set to 'Display Meridiem'
+              aria-label={intl.formatMessage({ id: 'Terra.timeInput.display.meridiem' })} // value in translations set to 'Display Meridiem'
               aria-readonly
               className={cx(['meridiem-display', { focused: this.state.meridiemFocused }])}
               onFocus={this.handleMeridiemInputFocus}
@@ -776,7 +774,7 @@ class TimeInput extends React.Component {
               key="meridiem_select_box"
             >
               <select
-                aria-label={this.context.intl.formatMessage({ id: 'Terra.timeInput.select.meridiem' })} // value in translations set to 'Select Meridiem'
+                aria-label={intl.formatMessage({ id: 'Terra.timeInput.select.meridiem' })} // value in translations set to 'Select Meridiem'
                 ref={(select) => { this.meridiemSelect = select; }}
                 onBlur={this.handleMeridiemBlur}
                 onFocus={this.handleMeridiemSelectFocus}
@@ -809,6 +807,5 @@ class TimeInput extends React.Component {
 
 TimeInput.propTypes = propTypes;
 TimeInput.defaultProps = defaultProps;
-TimeInput.contextTypes = contextTypes;
 
-export default TimeInput;
+export default injectIntl(TimeInput);