+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -6,9 +6,8 @@ Unreleased
 ### Changed
 * Screenshots updated for mobile view.
 * Update wdio test to work with changes in terra-date-picker
-
-### Changed
 * Renamed test and example SCSS files to end with `.module.scss` extension
+* Updated to interface with react-intl's Intl context rather than old React Context API
 
 4.4.0 - (June 19, 2019)
 ------------------